+++ packages/field/src/components/DatePicker/index.tsx
@@ -40,16 +40,13 @@ const FieldDatePicker: ProFieldFC<{
   const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
   const prefixCls = getPrefixCls('pro-field-date-picker');
   const [open, setOpen] = useState<boolean>(false);
+
   if (mode === 'read') {
-    const dom = (
-      <span ref={ref}>
-        {text ? moment(text).format(fieldProps.format || format || 'YYYY-MM-DD') : '-'}
-      </span>
-    );
+    const dom = text ? moment(text).format(fieldProps.format || format || 'YYYY-MM-DD') : '-';
     if (render) {
-      return render(text, { mode, ...fieldProps }, <span>{dom}</span>);
+      return render(text, { mode, ...fieldProps }, <>{dom}</>);
     }
-    return dom;
+    return <>{dom}</>;
   }
   if (mode === 'edit' || mode === 'update') {
     let dom;