+++ packages/layout/src/demos/base.tsx
@@ -133,6 +133,7 @@ export default () => {
       </ProLayout>
       <SettingDrawer
         pathname={pathname}
+        enableDarkTheme
         getContainer={() => document.getElementById('test-pro-layout')}
         settings={settings}
         onSettingChange={(changeSetting) => {