+++ packages/layout/src/utils/utils.ts
@@ -12,45 +12,24 @@ export const getOpenKeysFromMenuData = (menuData?: MenuDataItem[]) => {
     return pre;
   }, [] as string[]);
 };
-
 const themeConfig = {
-  daybreak: 'daybreak',
-  '#1890ff': 'daybreak',
-  '#F5222D': 'dust',
-  '#FA541C': 'volcano',
-  '#FAAD14': 'sunset',
-  '#13C2C2': 'cyan',
-  '#52C41A': 'green',
-  '#2F54EB': 'geekblue',
-  '#722ED1': 'purple',
+  daybreak: '#1890ff',
+  dust: '#F5222D',
+  volcano: '#FA541C',
+  sunset: '#FAAD14',
+  cyan: '#13C2C2',
+  green: '#52C41A',
+  geekblue: '#2F54EB',
+  purple: '#722ED1',
 };
-
-// eslint-disable-next-line @typescript-eslint/ban-types
-const invertKeyValues = (obj: Object) =>
-  Object.keys(obj).reduce((acc, key) => {
-    acc[obj[key]] = key;
-    return acc;
-  }, {});
-
-/**
- * #1890ff -> daybreak
- *
- * @param val
- */
-export function genThemeToString(val?: string): string {
-  return val && themeConfig[val] ? themeConfig[val] : undefined;
-}
-
 /**
  * Daybreak-> #1890ff
  *
  * @param val
  */
 export function genStringToTheme(val?: string): string {
-  const stringConfig = invertKeyValues(themeConfig);
-  return val && stringConfig[val] ? stringConfig[val] : val;
+  return val && themeConfig[val] ? themeConfig[val] : val;
 }
-
 export function clearMenuItem(menusData: MenuDataItem[]): MenuDataItem[] {
   return menusData
     .map((item) => {