+++ typings.d.ts
@@ -3,5 +3,3 @@ declare module '*.module.less';
 interface Window {
   DarkReader: any;
 }
-
-declare const DarkReader;