+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,58 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.17](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.16...@ant-design/pro-descriptions@1.10.17) (2021-12-17)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.16](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.13...@ant-design/pro-descriptions@1.10.16) (2021-12-13)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.13](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.12...@ant-design/pro-descriptions@1.10.13) (2021-12-10)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.12](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.11...@ant-design/pro-descriptions@1.10.12) (2021-12-10)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.11](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.10...@ant-design/pro-descriptions@1.10.11) (2021-12-09)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.10](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.9...@ant-design/pro-descriptions@1.10.10) (2021-12-08)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.9](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.8...@ant-design/pro-descriptions@1.10.9) (2021-12-07)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.8](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.7...@ant-design/pro-descriptions@1.10.8) (2021-12-01)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.6...@ant-design/pro-descriptions@1.10.7) (2021-12-01)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.5...@ant-design/pro-descriptions@1.10.6) (2021-11-30)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.4...@ant-design/pro-descriptions@1.10.5) (2021-11-23)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.3...@ant-design/pro-descriptions@1.10.4) (2021-11-23)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.10.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.2...@ant-design/pro-descriptions@1.10.3) (2021-11-22)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [1.10.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.10.1...@ant-design/pro-descriptions@1.10.2) (2021-11-18)
 
 ### Bug Fixes